requirements.txt
matplotlib==3.1.2
numpy==1.18.1
scikit-learn==0.21.3
scikit-image==0.16.2
pillow==7.0.0
scipy==1.4.1
tqdm==4.36.1
pybind11==2.4.3
torchvision==0.4.1
